+++ src/ui_event/spider_base_ui.py
@@ -10,7 +10,7 @@ from loguru import logger
 from ui_event.base_event import auto_start_spider_image, stop_spider_image, stop_download_image, edit_config_msg, \
     visit_web, about_message_lookup, online_look_image, performance_monitor, auto_play_image, \
     log_analyze_ui, face_detect_action, convert_folder_name, encoding_tools_convert, detect_installed_flag, \
-    user_upload_image, user_download_image, unzip_file_method, add_keyword_alert
+    user_upload_image, user_download_image, unzip_file_method, add_keyword_alert, img_category_ana
 
 
 @logger.catch
@@ -63,16 +63,19 @@ def base_menu(self):
     self.performance_menu.addAction(self.look_performance)
 
     self.tools_menu = QMenu('工具', self.menu_bar)
+    self.img_ana = QMenu('图片分析', self.tools_menu)
     self.log_analyze = QAction('日志分析', self.tools_menu)
     self.folder_name = QAction('名称转换', self.tools_menu)
     self.encoding_tools = QAction('编码转换', self.tools_menu)
     self.test_driver = QAction('驱动检测', self.tools_menu)
     self.unzip_file_menu = QAction('解压文件', self.tools_menu)
+    self.img_ana.triggered.connect(lambda: img_category_ana())
     self.log_analyze.triggered.connect(lambda: log_analyze_ui())
     self.folder_name.triggered.connect(lambda: convert_folder_name())
     self.encoding_tools.triggered.connect(lambda: encoding_tools_convert())
     self.test_driver.triggered.connect(lambda:  detect_installed_flag())
     self.unzip_file_menu.triggered.connect(lambda: unzip_file_method())
+    self.tools_menu.addMenu(self.img_ana)
     self.tools_menu.addAction(self.log_analyze)
     self.tools_menu.addAction(self.folder_name)
     self.tools_menu.addAction(self.encoding_tools)